Put everything into functions.
