+++ app/(authenticated)/[username]/page.tsx
@@ -2,181 +2,45 @@ import { IoArrowBackOutline } from "react-icons/io5";
 import { cookies } from "next/headers";
 import { notFound } from "next/navigation";
 import { redirect } from "next/navigation";
-import { sql, eq, desc } from "drizzle-orm";
 
 import { MdDateRange } from "react-icons/md";
 
 import { createClient } from "@/lib/supabase/server";
-import { db } from "@/lib/db";
-import * as schemas from "@/lib/db/schema";
+import { getUserDataByID, getUserSessionID, getUserDataByUsername } from "@/lib/getUserData";
+import getUserPosts from "@/lib/getUserPosts";
 
 import ProfileDashboardComponent from "@/components/Profile/ProfileDashboardComponent";
 import IconHeaderComponent from "@/components/Profile/IconHeaderComponent";
 
 import type { UserData } from "@/lib/types/userdata.types";
-
-async function getUserData(username: string): Promise<UserData | null> {
-    const cookieStore = cookies();
-    const supabase = await createClient(cookieStore);
-
-    const { data: userProfileData, error: userProfileDataError } = await supabase
-        .from("profile")
-        .select("*")
-        .eq("user_name", username);
-
-    if (userProfileDataError) {
-        console.error("[getUserData() Error]: ", userProfileDataError);
-        return null;
-    }
-
-    if (!userProfileData.length) {
-        const noUser: UserData = {
-            id: "",
-            userName: "",
-            displayName: "",
-            createdAt: new Date(),
-            avatarUrl: "",
-            headerUrl: "",
-            bio: ""
-        };
-
-        return noUser;
-    }
-
-    const userArray = userProfileData as any[];
-    const user = userArray[0];
-
-    const userProfile: UserData = {
-        id: user.id,
-        userName: user["user_name"],
-        displayName: user["display_name"],
-        createdAt: user["created_at"],
-        avatarUrl: user["avatar_url"],
-        headerUrl: user["header_url"],
-        bio: user.bio
-    }
-
-    return userProfile;
-}
-
-async function getCurrentUser(userID: string): Promise<UserData | null> {
-    const cookieStore = cookies();
-    const supabase = await createClient(cookieStore);
-
-    const { data: userProfileData, error: userProfileDataError } = await supabase
-        .from("profile")
-        .select("*")
-        .eq("id", userID);
-
-    if (userProfileDataError) {
-        console.error("[getUserData() Error]: ", userProfileDataError);
-        return null;
-    }
-
-    if (!userProfileData.length) {
-        const noUser: UserData = {
-            id: "",
-            userName: "",
-            displayName: "",
-            createdAt: new Date(),
-            avatarUrl: "",
-            headerUrl: "",
-            bio: ""
-        };
-
-        return noUser;
-    }
-
-    const userArray = userProfileData as any[];
-    const user = userArray[0];
-
-    const userProfile: UserData = {
-        id: user.id,
-        userName: user["user_name"],
-        displayName: user["display_name"],
-        createdAt: user["created_at"],
-        avatarUrl: user["avatar_url"],
-        headerUrl: user["header_url"],
-        bio: user.bio
-    }
-
-    return userProfile;
-}
-
-async function getUserPosts(username: string, userID: string) {
-    const getUserTweets = await db.select({
-        id: schemas.tweet.id,
-        authorInfo: {
-            authorDisplayName: sql<string>`${schemas.profile.displayName}`,
-            authorUserName: sql<string>`${schemas.profile.userName}`,
-            authorAvatarUrl: sql<string>`${schemas.profile.avatarUrl}`,
-            authorHeaderUrl: sql<string>`${schemas.profile.headerUrl}`,
-        },
-        textContent: schemas.tweet.postContent,
-        likeCount: sql<string>`COUNT(${schemas.like.id})`,
-        replyCount: sql<string>`COUNT(${schemas.reply.id})`,
-        createdAt: sql<string>`${schemas.tweet.createdAt}`,
-        hasLikedTweet: sql<boolean>`
-            EXISTS(
-                SELECT
-                    1
-                FROM
-                    ${schemas.like}
-                WHERE
-                    ${schemas.like.userID} = ${userID}
-                AND
-                    ${schemas.like.tweetID} = ${schemas.tweet.id}
-            )
-        `,
-        isReply: schemas.tweet.isReply
-    })
-        .from(schemas.tweet)
-        .leftJoin(schemas.profile, eq(schemas.tweet.profileID, schemas.profile.id))
-        .leftJoin(schemas.like, eq(schemas.tweet.id, schemas.like.tweetID))
-        .leftJoin(schemas.reply, eq(schemas.tweet.id, schemas.reply.replyTo))
-        .where(eq(schemas.profile.userName, username))
-        .orderBy(desc(schemas.tweet.createdAt))
-        .groupBy(schemas.tweet.id, schemas.profile.id)
-        .limit(50);
-
-    return getUserTweets;
-}
+import type { Tweet } from "@/lib/types/tweet.types";
 
 export default async function User({ params }: { params: { username: string } }) {
     const { username } = params;
     const cookieStore = cookies();
     const supabase = await createClient(cookieStore);
 
-    const userProfile = await getUserData(username);
+    const userProfile = await getUserDataByUsername(supabase, username);
 
     if (!userProfile) {
-        alert("[!!! ERROR !!!]: Server error check console.");
         return null;
     }
 
     if (!userProfile.id) notFound();
 
-    const { data, error: sessionErr } = await supabase.auth.getSession();
+    const userSession = await getUserSessionID(supabase);
 
-    if (sessionErr) {
-        console.error("[/[username] Session Error]: ", sessionErr);
-        alert("[!!! ERROR !!!]: Server error check console.");
+    if (!userSession) {
         return null;
     }
 
-    if (!data.session) redirect("/");
-
-    const currentUserID = data.session.user.id;
-    const currentUser = await getCurrentUser(currentUserID);
+    const currentUser: UserData | null = await getUserDataByID(supabase, userSession.id);
 
     if (!currentUser) {
-        alert("[!!! ERROR !!!]: Server error check console.");
         return null;
     }
 
-    if (!currentUser.id) redirect("/");
-
-    const userPosts = await getUserPosts(username, currentUserID);
+    const userPosts: Tweet[] = await getUserPosts(username, currentUser.id);
 
     const userJoinedDate = new Date(userProfile.createdAt);
     const userJoinedYear = userJoinedDate.getFullYear();
